refactor: 통계로직을 리뷰와 분리
