+++ backend/src/main/java/com/woowacourse/f12/domain/review/ReviewRepositoryCustomImpl.java
@@ -9,7 +9,6 @@ import static com.woowacourse.f12.support.RepositorySupport.toSlice;
 import com.querydsl.jpa.impl.JPAQuery;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import java.util.Collections;
-import java.util.List;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
 import org.springframework.data.domain.SliceImpl;
@@ -43,33 +42,5 @@ public class ReviewRepositoryCustomImpl implements ReviewRepositoryCustom {
                 .fetchJoin()
                 .orderBy(makeOrderSpecifiers(review, pageable));
         return new SliceImpl<>(query.fetch(), pageable, reviewIds.hasNext());
-    }
-
-    @Override
-    public List<CareerLevelCount> findCareerLevelCountByProductId(final Long productId) {
-        return jpaQueryFactory.select(new QCareerLevelCount(member.careerLevel, member.id.count()))
-                .from(review)
-                .innerJoin(review.member, member)
-                .where(
-                        review.product.id.eq(productId),
-                        member.careerLevel.isNotNull(),
-                        member.jobType.isNotNull()
-                )
-                .groupBy(member.careerLevel)
-                .fetch();
-    }
-
-    @Override
-    public List<JobTypeCount> findJobTypeCountByProductId(final Long productId) {
-        return jpaQueryFactory.select(new QJobTypeCount(member.jobType, member.id.count()))
-                .from(review)
-                .innerJoin(review.member, member)
-                .where(
-                        review.product.id.eq(productId),
-                        member.careerLevel.isNotNull(),
-                        member.jobType.isNotNull()
-                )
-                .groupBy(member.jobType)
-                .fetch();
     }
 }
