+++ backend/src/main/java/com/woowacourse/f12/domain/statistics/Countable.java
@@ -1,4 +1,4 @@
-package com.woowacourse.f12.domain.review;
+package com.woowacourse.f12.domain.statistics;
 
 import com.woowacourse.f12.domain.member.MemberInfo;
