+++ backend/src/test/java/com/woowacourse/f12/domain/review/ReviewRepositoryTest.java
@@ -1,16 +1,10 @@
 package com.woowacourse.f12.domain.review;
 
-import static com.woowacourse.f12.domain.member.CareerLevel.JUNIOR;
-import static com.woowacourse.f12.domain.member.CareerLevel.SENIOR;
-import static com.woowacourse.f12.domain.member.JobType.BACKEND;
-import static com.woowacourse.f12.domain.member.JobType.MOBILE;
 import static com.woowacourse.f12.support.fixture.MemberFixture.CORINNE;
 import static com.woowacourse.f12.support.fixture.MemberFixture.MINCHO;
-import static com.woowacourse.f12.support.fixture.MemberFixture.NOT_ADDITIONAL_INFO;
 import static com.woowacourse.f12.support.fixture.ProductFixture.KEYBOARD_1;
 import static com.woowacourse.f12.support.fixture.ProductFixture.KEYBOARD_2;
 import static com.woowacourse.f12.support.fixture.ReviewFixture.REVIEW_RATING_1;
-import static com.woowacourse.f12.support.fixture.ReviewFixture.REVIEW_RATING_2;
 import static com.woowacourse.f12.support.fixture.ReviewFixture.REVIEW_RATING_4;
 import static com.woowacourse.f12.support.fixture.ReviewFixture.REVIEW_RATING_5;
 import static org.assertj.core.api.Assertions.assertThat;
@@ -23,7 +17,6 @@ import com.woowacourse.f12.domain.member.Member;
 import com.woowacourse.f12.domain.member.MemberRepository;
 import com.woowacourse.f12.domain.product.Product;
 import com.woowacourse.f12.domain.product.ProductRepository;
-import java.util.List;
 import java.util.Optional;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
@@ -154,104 +147,6 @@ class ReviewRepositoryTest {
 
         // then
         assertThat(actual).isTrue();
-    }
-
-    @Test
-    void 제품에_대한_사용자_연차의_총_개수를_반환한다() {
-        // given
-        Product product = 제품_저장(KEYBOARD_1.생성());
-
-        Member corinne = CORINNE.추가정보를_입력하여_생성(SENIOR, MOBILE);
-        corinne = memberRepository.save(corinne);
-
-        Member mincho = MINCHO.추가정보를_입력하여_생성(JUNIOR, BACKEND);
-        mincho = memberRepository.save(mincho);
-
-        리뷰_저장(REVIEW_RATING_2.작성(product, corinne));
-        리뷰_저장(REVIEW_RATING_1.작성(product, mincho));
-
-        // when
-        List<CareerLevelCount> careerLevelCounts = reviewRepository.findCareerLevelCountByProductId(
-                product.getId());
-
-        // then
-        assertThat(careerLevelCounts).usingRecursiveFieldByFieldElementComparator()
-                .hasSize(2)
-                .containsOnly(
-                        new CareerLevelCount(JUNIOR, 1),
-                        new CareerLevelCount(SENIOR, 1)
-                );
-    }
-
-    @Test
-    void 제품에_대한_사용자_연차의_총_개수를_반환할때_추가정보가_없는_회원은_포함되지_않는다() {
-        // given
-        Product product = 제품_저장(KEYBOARD_1.생성());
-
-        Member corinne = CORINNE.생성();
-        corinne = memberRepository.save(corinne);
-        Member notAdditionalInfo = NOT_ADDITIONAL_INFO.생성();
-        notAdditionalInfo = memberRepository.save(notAdditionalInfo);
-
-        리뷰_저장(REVIEW_RATING_2.작성(product, corinne));
-        리뷰_저장(REVIEW_RATING_1.작성(product, notAdditionalInfo));
-
-        // when
-        List<CareerLevelCount> careerLevelCounts = reviewRepository.findCareerLevelCountByProductId(
-                product.getId());
-
-        // then
-        assertThat(careerLevelCounts).usingRecursiveFieldByFieldElementComparator()
-                .hasSize(1)
-                .containsOnly(new CareerLevelCount(SENIOR, 1));
-    }
-
-    @Test
-    void 제품에_대한_사용자_직군의_총_개수를_반환한다() {
-        // given
-        Product product = 제품_저장(KEYBOARD_1.생성());
-
-        Member corinne = CORINNE.추가정보를_입력하여_생성(SENIOR, MOBILE);
-        corinne = memberRepository.save(corinne);
-
-        Member mincho = MINCHO.추가정보를_입력하여_생성(JUNIOR, BACKEND);
-        mincho = memberRepository.save(mincho);
-
-        리뷰_저장(REVIEW_RATING_2.작성(product, corinne));
-        리뷰_저장(REVIEW_RATING_1.작성(product, mincho));
-
-        // when
-        List<JobTypeCount> jobTypeCounts = reviewRepository.findJobTypeCountByProductId(product.getId());
-
-        // then
-        assertThat(jobTypeCounts).usingRecursiveFieldByFieldElementComparator()
-                .hasSize(2)
-                .containsOnly(
-                        new JobTypeCount(MOBILE, 1),
-                        new JobTypeCount(BACKEND, 1)
-                );
-    }
-
-    @Test
-    void 제품에_대한_사용자_직군의_총_개수를_반환할때_추가정보가_없는_회원은_포함되지_않는다() {
-        // given
-        Product product = 제품_저장(KEYBOARD_1.생성());
-
-        Member corinne = CORINNE.생성();
-        corinne = memberRepository.save(corinne);
-        Member notAdditionalInfo = NOT_ADDITIONAL_INFO.생성();
-        notAdditionalInfo = memberRepository.save(notAdditionalInfo);
-
-        리뷰_저장(REVIEW_RATING_2.작성(product, corinne));
-        리뷰_저장(REVIEW_RATING_1.작성(product, notAdditionalInfo));
-
-        // when
-        List<JobTypeCount> jobTypeCounts = reviewRepository.findJobTypeCountByProductId(product.getId());
-
-        // then
-        assertThat(jobTypeCounts).usingRecursiveFieldByFieldElementComparator()
-                .hasSize(1)
-                .containsOnly(new JobTypeCount(BACKEND, 1));
     }
 
     @Test
